urgs-rag/app/routers/sql2text.py
from fastapi import APIRouter
from pydantic import BaseModel

router = APIRouter()

class SqlExplainRequest(BaseModel):
    sql: str

@router.post("/explain")
async def explain_sql(request: SqlExplainRequest):
    """
    Explain a SQL query using RAG context.
    """
    # 1. Retrieve relevant context (e.g., table schemas, lineage)
    # docs = vector_store_service.similarity_search(request.sql, k=2)
    
    # 2. Construct Prompt
    # 3. Call LLM
    
    return {"explanation": "This is a placeholder explanation for the SQL."}
