urgs-rag/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic-settings>=2.1.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface
langchain-chroma
chromadb
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers
tiktoken>=0.5.2
neo4j>=5.15.0
pypdf>=3.17.0
pymupdf>=1.24.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
openai>=1.10.0
rank_bm25>=0.2.2
tabulate
langchain-text-splitters
numpy<2
rarfile
docx2txt
python-multipart
rapidocr_onnxruntime>=1.4.0
opencv-python-headless
Pillow

